requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
cryptography>=42.0.0
httpx>=0.27.0
anthropic>=0.40.0
python-jose[cryptography]>=3.3.0
beautifulsoup4>=4.12.0
